requirements.txt
websockets>=12.0
httpx>=0.27.0
sounddevice>=0.5.0
numpy>=1.26.0
python-dotenv>=1.0.1
openai>=1.40.0
elevenlabs>=1.0.0
